fit: 调整数据类型

+++ home_work_1/cython/v2/home_work.pyx
@@ -5,7 +5,7 @@ cimport numpy as np
 cpdef target_mean(data,y_name,x_name):
     cdef long nrow = data.shape[0]
     cdef np.ndarray[double] result = np.asfortranarray(np.zeros(nrow),dtype=np.float64)
-    cdef np.ndarray[double] y = np.asfortranarray(data[y_name],dtype=np.float64)
+    cdef np.ndarray[int] y = np.asfortranarray(data[y_name],dtype=np.int32)
     cdef np.ndarray[double] x = np.asfortranarray(data[x_name],dtype=np.float64)
 
     target_mean_impl(result, y, x, nrow)
